+++ restli-common/src/main/java/com/linkedin/restli/common/RestConstants.java
@@ -59,7 +59,7 @@ public interface RestConstants
   String CONTENT_TYPE_PARAM_SYMBOL_TABLE = "symbol-table";
   String HEADER_CONTENT_ID = "Content-ID";
   String HEADER_SERVICE_SCOPED_PATH = "x-restli-service-scoped-path";
-  String HEADER_FETCH_SYMBOL_TABLE = "X-Restli-Symbol-Table-Request";
+  String HEADER_FETCH_SYMBOL_TABLE = "x-restli-symbol-table-request";
 
   // Default supported mime types.
   Set<String> SUPPORTED_MIME_TYPES = new LinkedHashSet<>(